+++ src/Phan/AST/ContextNode.php
@@ -607,6 +607,7 @@ class ContextNode
                 $this->emitIssue(
                     $custom_issue_type ?? ($expected_type_categories === self::CLASS_LIST_ACCEPT_OBJECT_OR_CLASS_NAME ? Issue::TypeExpectedObjectOrClassName : Issue::TypeExpectedObject),
                     $this->node->lineno ?? $this->context->getLineNumberStart(),
+                    ASTReverter::toShortString($this->node),
                     (string)$union_type->asNonLiteralType()
                 );
             } elseif ($expected_type_categories === self::CLASS_LIST_ACCEPT_OBJECT_OR_CLASS_NAME) {