+++ tests/files/src/0857_redefined_class_self.php
@@ -1,0 +1,14 @@
+<?php
+class X {
+    const DEFAULT_VALUE = 'x';
+    public function f1() {
+        return self::DEFAULT_VALUE;
+    }
+}
+
+class X {
+    const DEFAULT_VALUE2 = 'x';
+    public function f2() {
+        return self::DEFAULT_VALUE2;
+    }
+}