+++ src/Phan/Language/Type.php
@@ -147,13 +147,13 @@ class Type
      */
     public const type_regex =
         '('
-        . '(?:\??\((?-1)(?:\|(?-1))*\)|'  // Recursion: "?(T)" or "(T)" with brackets. Also allow parsing (a|b) within brackets.
+        . '(?:\??\((?-1)(?:[|&](?-1))*\)|'  // Recursion: "?(T)" or "(T)" with brackets. Also allow parsing (a|b) within brackets.
         . '(?:'
           . '\??(?:\\\\?Closure|callable)(\((?:[^()]|(?-1))*\))'  // `Closure(...)` can have matching pairs of () inside `...`, recursively
           . '(?:\s*:\s*'  // optional return type, can be ":T" or ":(T1|T2)" or ": ?(T1|T2)"
             . '(?:'
               . self::simple_noncapturing_type_regex . '|'  // Forbid ambiguity in `Closure():int[]` by disallowing complex return types without '()'. Always parse that as `(Closure():int)[]`.
-              . '\((?-2)(?:\s*\|\s*(?-2))*\)'
+              . '\((?-2)(?:\s*[|&]\s*(?-2))*\)'
             . ')'
           . ')?'
         . ')|'
@@ -162,16 +162,16 @@ class Type
         . '(?:'
           . '<'
             . '('
-              . '(?-5)(?:\|(?-5))*'
+              . '(?-5)(?:[|&](?-5))*'
               . '(?:\s*,\s*'
-                . '(?-5)(?:\|(?-5))*'
+                . '(?-5)(?:[|&](?-5))*'
               . ')*'
             . ')'
           . '>'
           . '|'
           . '\{('  // Expect either '{' or '<', after a word token.
-            . '(?:' . self::shape_key_regex . '\s*:)?\s*(?-6)(?:\|(?-6))*=?'  // {shape_key_regex:<type_regex>}
-            . '(?:,(?:\s*' . self::shape_key_regex . '\s*:)?\s*(?-6)(?:\|(?-6))*=?)*'  // {shape_key_regex:<type_regex>}
+            . '(?:' . self::shape_key_regex . '\s*:)?\s*(?-6)(?:[|&](?-6))*=?'  // {shape_key_regex:<type_regex>}
+            . '(?:,(?:\s*' . self::shape_key_regex . '\s*:)?\s*(?-6)(?:[|&](?-6))*=?)*'  // {shape_key_regex:<type_regex>}
           . ')?\})?'
         . ')'
         . '(?:\[\])*'
@@ -189,13 +189,13 @@ class Type
         '('
         . '('
           . '(?:'
-            . '\??\((?-1)(?:\|(?-1))*\)|'  // Recursion: "?(T)" or "(T)" with brackets. Also allow parsing (a|b) within brackets.
+            . '\??\((?-1)(?:[|&](?-1))*\)|'  // Recursion: "?(T)" or "(T)" with brackets. Also allow parsing (a|b) within brackets.
             . '(?:'
               . '\??(?:\\\\?Closure|callable)(\((?:[^()]|(?-1))*\))'  // `Closure(...)` can have matching pairs of () inside `...`, recursively
               . '(?:\s*:\s*'  // optional return type, can be ":T" or ":(T1|T2)"
                 . '(?:'
                   . self::simple_noncapturing_type_regex . '|'  // Forbid ambiguity in `Closure():int[]` by disallowing complex return types without '()'. Always parse that as `(Closure():int)[]`.
-                  . '\((?-2)(?:\s*\|\s*(?-2))*\)'  // Complicated return types can be placed within ().
+                  . '\((?-2)(?:\s*[|&]\s*(?-2))*\)'  // Complicated return types can be placed within ().
                 . ')'
               . ')?'
             . ')|'
@@ -203,16 +203,16 @@ class Type
             . '(' . self::simple_type_regex_or_this . ')'  // 3 patterns
             . '(?:<'
               . '('
-                . '(?-6)(?:\|(?-6))*'  // We use relative references instead of named references so that more than one one type_regex can be used in a regex.
+                . '(?-6)(?:[|&](?-6))*'  // We use relative references instead of named references so that more than one one type_regex can be used in a regex.
                 . '(?:\s*,\s*'
-                  . '(?-6)(?:\|(?-6))*'
+                  . '(?-6)(?:[|&](?-6))*'
                 . ')*'
               . ')'
               . '>'
               . '|'
               . '(\{)('  // Expect either '{' or '<', after a word token. Match '{' to disambiguate 'array{}'
-                . '(?:' . self::shape_key_regex . '\s*:)?\s*(?-8)(?:\|(?-8))*=?'  // {shape_key_regex:<type_regex>}
-                . '(?:,(?:\s*' . self::shape_key_regex . '\s*:)?\s*(?-8)(?:\|(?-8))*=?)*'  // {shape_key_regex:<type_regex>}
+                . '(?:' . self::shape_key_regex . '\s*:)?\s*(?-8)(?:[|&](?-8))*=?'  // {shape_key_regex:<type_regex>}
+                . '(?:,(?:\s*' . self::shape_key_regex . '\s*:)?\s*(?-8)(?:[|&](?-8))*=?)*'  // {shape_key_regex:<type_regex>}
               . ')?\})?'
             . ')'
           . '(?:\[\])*'