+++ src/Phan/Language/UnionType.php
@@ -90,7 +90,7 @@ class UnionType implements Serializable
     // phpcs:ignore Generic.NamingConventions.UpperCaseConstantName
     public const union_type_regex =
         Type::type_regex
-        . '(\s*\|\s*' . Type::type_regex . ')*';
+        . '(\s*[|&]\s*' . Type::type_regex . ')*';
 
     /**
      * @var string
@@ -103,7 +103,7 @@ class UnionType implements Serializable
     // phpcs:ignore Generic.NamingConventions.UpperCaseConstantName
     public const union_type_regex_or_this =
         Type::type_regex_or_this
-        . '(\s*\|\s*' . Type::type_regex_or_this . ')*';
+        . '(\s*[|&]\s*' . Type::type_regex_or_this . ')*';
 
     /**
      * @var list<Type> This is an immutable list of unique types.
@@ -501,7 +501,7 @@ class UnionType implements Serializable
     private static function extractTypeParts(string $type_string): array
     {
         $parts = [];
-        foreach (\explode('|', $type_string) as $part) {
+        foreach (\preg_split('@[|&]@', $type_string) as $part) {
             $parts[] = \trim($part);
         }