+++ NEWS.md
@@ -6,6 +6,7 @@ Phan NEWS
 New Features(Analysis):
 + Emit `PhanUndeclaredConstantOfClass` (severity critical) for undeclared class constants instead of `PhanUndeclaredConstant` (severity normal)
   This should not be confused with `PhanUndeclaredClassConstant`, which already exists and refers to accessing class constants of classes that don't exist.
++ Emit the expression that's an invalid object with issue types such as `PhanTypeExpectedObject*`, `PhanTypeInvalidInstanceof` (#3717)
 
 Bug fixes:
 + Fix incorrect file location for other definition in `PhanRedefinedClassReference` under some circumstances.